pyproject.toml
[project]
name = "supply-sense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.68",
    "boto3-stubs>=1.42.68",
    "fastapi[standard]>=0.135.1",
    "google-genai>=1.67.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langgraph>=1.0.10",
    "pathlib>=1.0.1",
    "psycopg2>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "tavily-python>=0.7.23",
    "types-psycopg2>=2.9.21.20260223",
    "uvicorn>=0.41.0",
]

[tool.fastapi]
entrypoint = "backend.main:app"
